refactor: clean up ESLint configuration and improve code readability
